+++ StockMarketSimulator.Wallets.Kernel/Services/IWalletsService.cs
@@ -6,8 +6,8 @@ namespace StockMarketSimulator.Wallets.Kernel.Services
     public interface IWalletsService
     {
         Task<WalletDto> Get(Guid walletId);
-        Task Deposit(Guid walletId, string stockSymbol, float quantity);
-        Task Withdraw(Guid walletId, string stockSymbol, float quantity);
-        Task Exchange(Guid walletId, string baseSymbol, string quoteSymbol, float quantity);
+        Task Deposit(Guid walletId, string stockSymbol, double quantity);
+        Task Withdraw(Guid walletId, string stockSymbol, double quantity);
+        Task Exchange(Guid walletId, string baseSymbol, string quoteSymbol, double quantity);
     }
 }
